pyproject.toml
[project]
name = "fall-watch"
version = "0.1.0"
description = "Real-time fall detection via YOLOv8 pose estimation and Telegram alerts"
requires-python = ">=3.13"
dependencies = [
    "ultralytics>=8.3",
    "opencv-python-headless>=4.10",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[project.scripts]
fall-watch = "fall_watch.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fall_watch"]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8.3",
    "types-requests>=2.32",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "RUF", # ruff-specific
]
ignore = [
    "E501", # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["fall_watch"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
